Set process group in subprocess.Popen for rosbag recording
This commit modifies the subprocess.Popen call used for starting the rosbag recording process. By adding the preexec_fn=os.setpgrp argument, the rosbag process is now spawned in a new process group. This change ensures that the rosbag recording process can be managed more effectively, particularly in scenarios where it needs to be terminated independently of the parent process.