fix: rotate left thumb tip mesh to compensate for model quaternion
The left thumb tip (l_f_link1_4) had incorrect orientation due to a quaternion
rotation applied in the model XML files. Instead of modifying all model files,
this fix pre-rotates the STL mesh by 180 degrees around the X-axis to compensate
for the quat="0.000796327 1 0 0" rotation applied in the models.

This ensures the thumb tip appears correctly oriented when loaded in MuJoCo.